fix: cleanning stuff
