+++ frontend/src/utils.ts
@@ -12,7 +12,6 @@ export function getCookie(name: string): string | null {
   if (parts.length === 2) {
     return parts.pop()?.split(";").shift() ?? null;
   }
-
   return null;
 }
